Refactor into seperate content factories
