+++ src/Contracts/Factory.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Aerni\Factory\Contracts;
+
+interface Factory
+{
+    public function run(): void;
+}
