+++ src/Utils.php
@@ -7,15 +7,15 @@ class Utils
     /**
      * Recursively map an array to a callback function.
      *
-     * @param  function  $callback
+     * @param function $callback
      */
-    public static function array_map_recursive(array $array, $callback): array
+    public static function mapRecursive(array $array, $callback): array
     {
         $output = [];
 
         foreach ($array as $key => $value) {
             if (is_array($value)) {
-                $output[$key] = self::array_map_recursive($value, $callback);
+                $output[$key] = self::mapRecursive($value, $callback);
             } else {
                 $output[$key] = $callback($value, $key);
             }
